% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproofer_install.R
\name{dataproofer_install}
\alias{dataproofer_install}
\title{Install dataproofer}
\usage{
dataproofer_install(force = F)
}
\arguments{
\item{force}{should the user be prompted to install?}
}
\value{
An installed lib
}
\description{
Install the underlying dataproofer npm package.
}
