% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_available.R
\name{node_available}
\alias{node_available}
\title{Check npm and Node status}
\usage{
node_available()
}
\arguments{
\item{force}{should the user be prompted to install?}
}
\value{
Warning if not available, message if available.
}
\description{
Are npm and Node available on this device?
}
