% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proof.R
\name{proof}
\alias{proof}
\title{Proof tabular data}
\usage{
proof(
  data,
  tests = "all",
  out = F,
  out_format = "json",
  summary = F,
  verbose = F,
  exclude = F
)
}
\arguments{
\item{data}{Tibble or dataframe to test.}

\item{tests}{Which tests do you want to perform? Defaults to "all", other options include "core", "info", "stats" and "geo".}

\item{out}{Export the results to a json file. Specify a path to export.}

\item{out_format}{Choose between "json" (default) and "json-pretty".}

\item{summary}{Exclude tests that passed.}

\item{verbose}{Include descriptions about each column.}
}
\description{
Specify a tibble or dataframe and perform automated data proofing tests.
}
