% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proof_custom.R
\name{proof_custom}
\alias{proof_custom}
\title{Proof tabular data}
\usage{
proof_custom(data, options = "")
}
\arguments{
\item{data}{Tibble or dataframe to test.}

\item{options}{Pass on arguments directly to the Node package.}
}
\description{
Proof tabular data.
}
